+++ app/Http/Controllers/API/AuthController.php
@@ -1,4 +1,10 @@
 <?php
+/**
+ * ユーザー登録・accessToken発行、ログイン、ログアウト
+ * 
+ * @version 1.0
+ * @author 米内
+ */
 
 namespace App\Http\Controllers\API;
 
@@ -10,6 +16,13 @@ use Illuminate\Http\Request;
 
 class AuthController extends Controller
 {
+    /**
+     * ユーザー登録処理
+     * 認証用のaccessTokenを発行
+     * 
+     * @param \Illuminate\Http\Request
+     * @return string|string 
+     */
     public function register(Request $request)
     {
         $validatedData = $request->validate([
@@ -27,6 +40,14 @@ class AuthController extends Controller
 
         return response(['user' => $user, 'access_token' => $accessToken]);
     }
+
+    /**
+     * ログイン認証処理
+     * SessionStrageにaccessTokenを保存
+     * 
+     * @param \Illuminate\Http\Request
+     * @return string|mixed 認証情報の保存、ログインユーザー情報のレスポンス
+     */
     public function login(Request $request)
     {
         $loginData = $request->validate([
@@ -43,6 +64,13 @@ class AuthController extends Controller
         return response(['user' => auth()->user(), 'access_token' => $accessToken]);
     }
 
+    /**
+     * ログアウト
+     * SessionStrageのaccessTokenを削除
+     * 
+     * @param \Illuminate\Http\Request
+     * @return string|string ログアウト完了メッセージ
+     */
     public function logout(Request $request)
     {
         $request->user()->token()->revoke();
