+++ app/Http/Controllers/API/CollectionController.php
@@ -1,4 +1,11 @@
 <?php
+/**
+ * POPUP表示グループ情報を登録、削除処理
+ * 
+ * 
+ * @version 1.0
+ * @author 米内
+ */
 
 namespace App\Http\Controllers\API;
 
@@ -11,12 +18,23 @@ use Illuminate\Support\Facades\Validator;
 
 class CollectionController extends Controller
 {
+    /**
+     * グループ一覧取得
+     * 
+     * @return \Illuminate\Http\Response
+     */
     public function index()
     {
         $collections = Collection::all();
         return response([ 'collections' => $collections, 'message' => 'Retrieved successfully'], 200);
     }
 
+    /**
+     * グループ登録
+     * 
+     * @param \Illuminate\Http\Request
+     * @return \Illuminate\Http\Response
+     */
     public function store(Request $request)
     {
         $data = $request->all();
@@ -34,6 +52,12 @@ class CollectionController extends Controller
         return response([ 'collection' => $collection, 'message' => 'Created successfully'], 200);
     }
 
+    /**
+     * グループ削除
+     * 
+     * @param \Models\Model\Collection $collection
+     * @return \Illuminate\Http\Response
+     */
     public function destroy(Collection $collection)
     {
         $collection->delete();
