+++ app/Http/Controllers/API/NotificationController.php
@@ -1,43 +1,63 @@
 <?php
+/**
+ * POPUP情報を登録、表示、削除するための指令を行う
+ * 
+ * @version 1.0
+ * @author 米内
+ */
 
 namespace App\Http\Controllers\API;
 
-use App\Http\Controllers\Controller;
-use App\Models\Notification;
-use App\Models\User;
-use App\Models\NotificationUser;
 use Illuminate\Http\Request;
-use App\Http\Resources\NotificationResource;
 use Illuminate\Support\Facades\Validator;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Auth;
+use App\Http\Controllers\Controller;
+use App\Http\Resources\NotificationResource;
+use App\Models\Notification;
+use App\Models\NotificationUser;
+use App\Models\User;
 
 class NotificationController extends Controller
 {
+    /**
+     * ユーザーに表示するPOPUP情報取得
+     * リクエストユーザーIDを参照し、表示すべきPOPUP情報のみ取得する
+     * 
+     * @return \Illuminate\Http\Response
+     */
     public function notify() {
-        $notification = new NotificationUser();
-        $notify_user = $notification->check_notify_user();
-        $auth_user = $notification->auth_user();
-         
+        $notification = new Notification();
+        $user = new User();
+        
         if(Auth::check()) {
-                return response([ 'notifies' => $notify_user, 'user_id' => $auth_user, 'message' => 'POPUP表示!!!'], 200);
+            $auth_user_id = $user->auth_user_id();
+            $notifies = $notification->check_notify_user($auth_user_id);
+            
+            return response([ 'notifies' => $notifies, 'user_id' => $auth_user_id, 'message' => 'POPUP表示!!!'], 200);
         } else {
             return response( '未認証' );
         }
     }
 
     /**
-     * Display a listing of the resource.
-     *
+     * POPUP非表示処理
+     * ユーザーが「次回から表示しない」or「既読」にチェック&閉じた時に実行
+     * 
+     * @param \Illuminate\Http\Request
      * @return \Illuminate\Http\Response
      */
-    public function index()
+    public function hidepopup(Request $request)
     {
-        //
+        $data = $request->all();
+
+        $hidepopup = NotificationUser::create($data);
+
+        return response([ 'hidepopup' => $hidepopup, 'message' => 'Created successfully'], 200);
     }
 
     /**
-     * Store a newly created resource in storage.
+     * POPUP情報を登録
      *
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
@@ -45,7 +65,7 @@ class NotificationController extends Controller
     public function store(Request $request)
     {
         $data = $request->all();
-
+        $notification = Notification::create($data);
         $validator = Validator::make($data, [
             'start_date' => 'required',
             'end_date' => 'required',
@@ -56,48 +76,24 @@ class NotificationController extends Controller
             'group_id' => 'required',
             'collection_id' => 'required',
         ]);
-
         
-        if($validator->fails()){
+        if($validator->fails()) {
             return response(['error' => $validator->errors(), 'Validation Error']);
         }
-
-        $notification = Notification::create($data);
 
         return response([ 'notification' => new NotificationResource($notification), 'message' => 'Created successfully'], 200);
     }
 
     /**
-     * Display the specified resource.
-     *
-     * @param  \App\Models\Notification  $notification
-     * @return \Illuminate\Http\Response
-     */
-    public function show(Notification $notification)
-    {
-        return response([ 'notification' => new NotificationResource($notification), 'message' => 'Retrived Successfully'], 200);
-    }
-
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Models\Notification  $notification
-     * @return \Illuminate\Http\Response
-     */
-    public function update(Request $request, Notification $notification)
-    {
-        //
-    }
-
-    /**
-     * Remove the specified resource from storage.
+     * POPUP情報を削除
      *
      * @param  \App\Models\Notification  $notification
      * @return \Illuminate\Http\Response
      */
     public function destroy(Notification $notification)
     {
-        //
+        $notification->delete();
+
+        return response(['message', 'Deleted!']);
     }
 }
