+++ app/Http/Controllers/API/UserController.php
@@ -1,4 +1,10 @@
 <?php
+/**
+ * ユーザー情報を表示するための指令を行う
+ * 
+ * @version 1.0
+ * @author 米内
+ */
 
 namespace App\Http\Controllers\API;
 
@@ -12,17 +18,13 @@ use App\Models\NotificationUser;
 
 class UserController extends Controller
 {
+    /**
+     * ユーザー一覧を取得
+     *
+     * @return \Illuminate\Http\Response
+     */
     public function index() {
         $users = USER::all();
         return response([ 'users' => $users, 'message' => 'Retrieved successfully'], 200);
-    }
-
-    public function hidepopup(Request $request)
-    {
-        $data = $request->all();
-
-        $hidepopup = NotificationUser::create($data);
-
-        return response([ 'hidepopup' => $hidepopup, 'message' => 'Created successfully'], 200);
     }
 }
