+++ app/Models/NotificationUser.php
@@ -5,42 +5,33 @@ namespace App\Models;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\Pivot;
-use Illuminate\Support\Facades\DB;
-use Illuminate\Support\Facades\Auth;
 
+/**
+ * App\Models\NotificationUser
+ *
+ * @property int $id
+ * @property int $notification_id
+ * @property int $user_id
+ * @property int $read
+ * @property int $hide_next
+ * @property \Illuminate\Support\Carbon|null $created_at
+ * @property \Illuminate\Support\Carbon|null $updated_at
+ * @method static \Illuminate\Database\Eloquent\Builder|NotificationUser newModelQuery()
+ * @method static \Illuminate\Database\Eloquent\Builder|NotificationUser newQuery()
+ * @method static \Illuminate\Database\Eloquent\Builder|NotificationUser query()
+ * @method static \Illuminate\Database\Eloquent\Builder|NotificationUser whereCreatedAt($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|NotificationUser whereHideNext($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|NotificationUser whereId($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|NotificationUser whereNotificationId($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|NotificationUser whereRead($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|NotificationUser whereUpdatedAt($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|NotificationUser whereUserId($value)
+ */
 class NotificationUser extends Model
 {
     use HasFactory;
 
     protected $fillable = ['notification_id','user_id','read','hide_next','created_at','updated_at'];
+
     protected $table = 'notification_user';
-
-    public function check_notify_user() {
-        $query = DB::table('notifications as n')
-            ->select( 'n.id as id',
-                    'n.image as image',
-                    'n.already_read as already_read', 
-                    'n.hide_next_time as hide_next_time', 
-                    'n.jump_link as jump_link', 
-                    'n.notify_priority as notify_priority',
-                    'nu.read as read',
-                    'nu.hide_next as hide_next' )
-            ->leftJoin('notification_user as nu', function ($join) {
-                $join->on('n.id', "=", 'nu.notification_id')
-                ->where('nu.user_id', "=", auth()->user()->id);
-            })
-            ->where('start_date', "<=", date("Y-m-d H:i:s"))
-            ->where('end_date', ">=", date("Y-m-d H:i:s"))
-            ->where('nu.user_id', null)
-            ->orderBy('n.notify_priority')
-            ->orderBy('n.id')
-            ->get();
-
-        return $query;
-    }
-
-    public function auth_user() {
-        return Auth::user()->id;
-    }
-
 }
