+++ routes/api.php
@@ -11,7 +11,6 @@ use App\Http\Controllers\API\NotificationController;
 Route::apiResource('/collection', CollectionController::class);
 
 Route::get('/user', [UserController::class, 'index']);
-Route::post('/hidepopup', [UserController::class, 'hidepopup'])->middleware('auth:api');
 
 Route::post('/register', [AuthController::class, 'register']);
 Route::post('/login', [AuthController::class, 'login']);
@@ -21,3 +20,4 @@ Route::apiResource('/ceo', (CEOController::class))->middleware('auth:api');
 
 Route::apiResource('/notification', (NotificationController::class))->middleware('auth:api');
 Route::get('/notify', [NotificationController::class, 'notify'])->middleware('auth:api');
+Route::post('/hidepopup', [NotificationController::class, 'hidepopup'])->middleware('auth:api');
